requirements.txt
Flask==2.0.2
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Werkzeug==2.0.3
PyJWT==2.3.0
SQLAlchemy==1.4.27
marshmallow==3.19.0
